piControl: Support more LEDs for event trigger
Function revpi_led_trigger_event() uses two masks (one for the previous and
one for the current active LEDs) to decide for which LED to trigger an
event. Since both masks are only 8 bit wide, only 8 LED states can be
distinguished. This is not sufficient for the REVPI Flat. So change the
function parameters to take 16Bit wide masks instead of the 8 bit wide
ones.

+++ revpi_common.c
@@ -23,7 +23,7 @@
 
 #define VCMSG_ID_ARM_CLOCK 0x000000003	/* Clock/Voltage ID's */
 
-void revpi_led_trigger_event(u8 led_prev, u8 led)
+void revpi_led_trigger_event(u16 led_prev, u16 led)
 {
 	u8 changed = led_prev ^ led;
 	if (changed == 0)
